minimum cost for travelling
There are N stations on route of a train. The train goes from station 0
to N-1. The ticket cost for all pair of stations (i, j) is given where j
is greater than i. Find the minimum cost to reach the destination.

Consider the following example:

Input:
cost[N][N] = { {0, 15, 80, 90},
{INF, 0, 40, 50},
{INF, INF, 0, 70},
{INF, INF, INF, 0}
};